requirements.txt
Flask

beautifulsoup4
requests

pandas
matplotlib
statsmodels

python-dotenv
azure-ai-documentintelligence